Pre-empt pip list format change
Following discussion in docker-library/python#187 it was decided that
rather than refactor this check, it should be removed. This was because
the original error condition it was designed to prevent could not be
reproduced. This is done with the understanding that it may need to be
revisited if any recurrance (of docker-library/python#100) is
encountered in future.