Fix segmentation fault when building Python 3.6 image
This fixes the issues reported in #160 for the Alpine build. For Python
3.6, the build succeeds but test illustrate that `ctypes` related test
fail with segmentation faults. This also prevents optimized builds
because they rely on a full run of the test suite.

The fix is based on the release candidate build of Python 3.6.1 in
Alpine Edge. It switches to system libraries for `libffi` and `expat` in
the same way it's used in [the Alpine build
file](https://git.alpinelinux.org/cgit/aports/tree/main/python3/APKBUILD).