Include python-gdb.py in images shipped with debugging symbols
Starting with gdb 7, it's possible to extend gdb with Python. CPython
is shipped with a script that allows to extract high-level information
specific to the interpreter, such has application-level stack traces,
the values of local variables, etc (see https://devguide.python.org/gdb/
for details). Include this script in the images that are shipped with
debugging symbols.

The script is copied to /usr/share/gdb/auto-load/usr/local/bin/python${VERSION}-gdb.py
and loaded automatically when gdb is used to debug a process running the
python binary.

The image size increase is negligible (~72KiB). Users will have to
install gdb separately, if they need it.

Closes #89