add symlinks in /usr/bin
Using the stock python:slim container is really difficult if you already have an existing script with `#!/usr/bin/python3` on top. You will get a confusing error like:

```
/usr/bin/bash: line 140: ./check.py: cannot execute: required file not found
```

... while the script otherwise works perfectly well. I understand that it's relatively standard to install things in `/usr/local` and I generally support this, but in this very specific case, it seems extremely backwards.

I also understand that, normally, you should *install* a script with an `entrypoint` and so on, but we shouldn't let perfection be the enemy of good here.