Do not set PROFILE_TASK environment variable
When this is set, the `--enable-optimizations` option is essentially disabled,
as the profile task step is skipped.

---

Tianon's commit revising note: this also disables optimizations on Alpine entirely, as they fail many tests (which is a known issue upstream AND in Alpine), and the Alpine builds are intended to be optimized for disk size (not speed) anyhow.